+++ lib/modl/parser/throwing_error_listener.rb
@@ -29,7 +29,7 @@ module MODL::Parser
     include Singleton
 
     def syntax_error(_recognizer, _offending_symbol, _line, _char_position_in_line, _msg, _e)
-      raise Antlr4::Runtime::ParseCancellationException, 'line' + _line.to_s + ':' + _char_position_in_line.to_s + ' ' + _msg.to_s
+      raise Antlr4::Runtime::ParseCancellationException, 'line ' + _line.to_s + ':' + _char_position_in_line.to_s + ' ' + _msg.to_s
     end
 
     def report_ambiguity(_recognizer, _dfa, _start_index, _stop_index, _exact, _ambig_ilts, _configs)