+++ lib/modl/parser/parsed.rb
@@ -30,6 +30,7 @@ require 'modl/parser/file_importer'
 require 'antlr4/runtime/parse_cancellation_exception'
 require 'modl/parser/sutil'
 require 'modl/parser/modl_class'
+require 'modl/parser/modl_array'
 require 'modl/parser/modl_method'
 require 'modl/parser/modl_index'
 require 'modl/parser/modl_keylist'
@@ -211,6 +212,8 @@ module MODL
           @final = false
           @file_importer = FileImporter.instance
           @loaded = false
+          @array = nil
+          @map = nil
         end
 
         def find_property(key)
@@ -284,6 +287,7 @@ module MODL
           return if @type == 'import'
           return if @type == 'allow'
           return if @type == 'expect'
+          return if @type == 'array'
 
           {@key => @text}
         end
@@ -336,6 +340,8 @@ module MODL
           case @type
           when 'class'
             ClassExtractor.extract(self, @global)
+          when 'array'
+            ArrayExtractor.extract(self, @global)
           when 'id'
             extract_value
           when 'name'
@@ -478,6 +484,8 @@ module MODL
           @type = 'hidden' if @key.start_with? '_'
           @type = 'allow' if @key.downcase == '*allow'
           @type = 'expect' if @key.downcase == '*expect'
+          @type = 'of' if @key.downcase == '*of'
+          @type = 'array' if @key.downcase == '*array'
         end
       end