+++ lib/modl/parser/ref_processor.rb
@@ -22,7 +22,6 @@
 # OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 # THE SOFTWARE.
 
-require 'modl/parser/parsed'
 require 'punycode'
 require 'modl/parser/sutil'